change api endpoint file  createTournament to newTournament
