+++ app/interfaces/games/tournament/TournamentSettings.ts
@@ -1,0 +1,14 @@
+import { GameSettings } from "../lobby/GameSettings";
+
+type bracketTypeSettings =
+  | "singleElimination"
+  | "doubleElimination"
+  | "roundRobin"
+  | "swissRound";
+
+export interface TournamentSettings {
+  bracketType: bracketTypeSettings;
+  matchWinner: string;
+  players: string[];
+  gameData: GameSettings;
+}
